+++ imod/prepare/layer.py
@@ -4,7 +4,7 @@ This module contains all kinds of utilities to work with layers.
 
 from imod.typing import GridDataArray
 from imod.typing.grid import preserve_gridtype, zeros_like
-
+from imod.common.utilities.layer import create_layered_top
 
 @preserve_gridtype
 def get_upper_active_layer_number(active: GridDataArray) -> GridDataArray:
@@ -72,28 +72,3 @@ def get_lower_active_grid_cells(active: GridDataArray) -> GridDataArray:
     layer = active.coords["layer"]
     lower_active_layer = get_lower_active_layer_number(active)
     return layer == lower_active_layer
-
-
-def create_layered_top(bottom: GridDataArray, top: GridDataArray) -> GridDataArray:
-    """
-    Create a top array with a layer dimension, from a top array with no layer
-    dimension and a bottom array with a layer dimension. The (output) top of
-    layer n is assigned the bottom of layer n-1.
-
-    Parameters
-    ----------
-    bottom: {DataArray, UgridDataArray}
-        Bottoms with layer dimension
-    top: {DataArray, UgridDataArray}
-        Top, without layer dimension
-
-    Returns
-    -------
-    new_top: {DataArray, UgridDataArray}
-        Top with layer dimension.
-    """
-    new_top = zeros_like(bottom)
-    new_top[0] = top
-    new_top[1:] = bottom[0:-1].values
-
-    return new_top
